search: Migrate comby logic to internal/comby
In preparation for moving the responsibility of querying Zoekt for
structural search file matches to `searcher`, this moves comby-specific
logic from the frontend to `internal/comby` so it can be re-used in
`searcher`.
